Moving more files to AGAS module
- flyby: remove resolver_client.hpp
- flyby: fixing integral conversion warning

+++ examples/quickstart/hello_world_distributed.cpp
@@ -95,7 +95,8 @@ void hello_world_foreman()
             // task will actually run on that worker thread.
             hpx::parallel::execution::default_executor exec(
                 hpx::threads::thread_schedule_hint(
-                    hpx::threads::thread_schedule_hint_mode::thread, worker));
+                    hpx::threads::thread_schedule_hint_mode::thread,
+                    static_cast<std::int16_t>(worker)));
             futures.push_back(hpx::async(exec, hello_world_worker, worker));
         }
